changed some parameters from int to double. Ran the tests

+++ Lab-03/RideNow/src/Location.java
@@ -4,13 +4,13 @@ public class Location
     private double x;
     private double y;
 
-    public Location(String label, int x, int y){
+    public Location(String label, double x, double y){
         this.label=label;
         this.x=x;
         this.y=y;
 
     }
-    public Location(int x,int y){
+    public Location(double x,double y){
         this.label="Unknown";
         this.x=x;
         this.y=y;
